pyproject.toml
[project]
name = "scrtrans"
version = "0.1.0"
description = "Desktop overlay that captures a user-selected screen region and streams translations for subtitles using configurable AI services"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "hazem" }
]
keywords = ["translation", "subtitle", "overlay", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Multimedia :: Video",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
]

dependencies = [
    "PyQt5==5.15.11",
    "Pillow>=12.0.0",
    "pyautogui>=0.9.54",
    "openai>=2.8.1",
    "google-genai>=1.52.0",
    "imagehash>=4.3.2",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pywin32>=311",
    "winocr>=0.0.1",
    "onnxruntime>=1.23.2",
    "rapidocr-onnxruntime>=1.4.4",
    "python-dotenv>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "services"]

[project.scripts]
scrtrans = "main:main"

[project.urls]
Homepage = "https://github.com/hazemAI/Transgemi"
Repository = "https://github.com/hazemAI/Transgemi"
Issues = "https://github.com/hazemAI/Transgemi/issues"
